pyproject.toml
[project]
name = "ibkr-toolkit"
version = "0.1.0"
description = "A comprehensive toolkit for IBKR trading data processing and management"
authors = [
    {name = "IBKR Toolkit Contributors"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "xmltodict>=0.13.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-dotenv>=1.0.0",
    "ib-async>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
    "mypy>=1.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
ibkr-toolkit = "ibkr_toolkit.cli:main"
ibkr-toolkit-stop-loss = "ibkr_toolkit.stop_loss_cli:main"

[project.urls]
Homepage = "https://github.com/undefined443/ibkr-toolkit"
Repository = "https://github.com/undefined443/ibkr-toolkit"
Documentation = "https://github.com/undefined443/ibkr-toolkit/blob/main/README.md"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=ibkr_toolkit --cov-report=term-missing"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
